feat: the profile picture can change.
