+++ index.html
@@ -6,7 +6,7 @@
   <link rel="icon" type="image/svg+xml" href="/vite.svg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <link href="/src/style.css" rel="stylesheet">
-  <script defer type="module" src="/src/main.js"></script>
+  <script defer type="module" src="/src/profile.js"></script>
   <link rel='stylesheet' href='https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.7.2/css/all.css'
     integrity='sha512-GmZYQ9SKTnOea030Tbiat0Y+jhnYLIpsGAe6QTnToi8hI2nNbVMETHeK4wm4MuYMQdrc38x+sX77+kVD01eNsQ=='
     crossorigin='anonymous' />
@@ -25,42 +25,65 @@
     <!-- imp! left side nav -->
 
     <div class="container grid fixed grid-cols-5 min-h-dvh min-w-full ">
-      <div class="aside flex flex-col gap-10 h-full w-full border-r-1 border-gray-300 min-h-dvh text-white ...">
-        <h2 class="text-4xl font-semibold font-logoText2 ps-8 pb-8 mt-10 text-vibPink">Photo</h2>
+
+      <div
+        class="aside fixed bottom-0 right-0 left-0 z-30 flex  justify-evenly gap-5 h-20 w-full border-r-1 border-gray-300 bg-black text-white ... lg:hidden">
+        <h2 class="flex items-center text-4xl font-semibold font-logoText2 text-vibPink">Photo</h2>
         <div class="flex items-center hover:text-darkOrange"><i
-            class="fa-solid fa-house-user cursor-pointer text-xl ps-8"></i> <span
-            class="ps-4 cursor-pointer text-xl">Home</span></div>
-        <div class="flex items-center hover:text-darkOrange"> <i
-            class="fa-solid fa-user cursor-pointer text-xl ps-8"></i>
-          <span class="ps-4 cursor-pointer text-xl">Profile</span>
+            class="fa-solid fa-house-user cursor-pointer text-xl"></i> <span class="ps-4 cursor-pointer text-xl"></span>
         </div>
-        <!-- <div class="flex items-center hover:text-darkOrange"> <i
-            class="fa-solid fa-magnifying-glass cursor-pointer text-xl ps-8"></i> <span
-            class="ps-4 cursor-pointer text-xl">Search</span></div> -->
+        <div class="flex items-center hover:text-darkOrange"> <i class="fa-solid fa-user cursor-pointer text-xl"></i>
+          <span class="ps-4 cursor-pointer text-xl"></span>
+        </div>
 
         <div class="flex items-center hover:text-darkOrange"> <i
-            class="fa-solid fa-people-group cursor-pointer text-xl ps-8"></i> <span
-            class="ps-4 cursor-pointer text-xl">Explore</span> </div>
-        <div class="flex items-center hover:text-darkOrange"> <i
-            class="fa-solid fa-plus cursor-pointer text-xl ps-8"></i> <span
-            class="ps-4 cursor-pointer text-xl">Create</span>
+            class="fa-solid fa-people-group cursor-pointer text-xl"></i> <span
+            class="ps-4 cursor-pointer text-xl"></span> </div>
+        <div class="flex items-center hover:text-darkOrange"> <i class="fa-solid fa-plus cursor-pointer text-xl"></i>
+          <span class="ps-4 cursor-pointer text-xl"></span>
         </div>
-        <div class="flex-end items-center hover:text-darkOrange"> <i
-            class="fa-solid fa-bars cursor-pointer text-xl ps-8"></i> <span
-            class="ps-4 cursor-pointer text-xl">More</span>
+        <div class=" flex items-center hover:text-darkOrange"> <i class="fa-solid fa-bars cursor-pointer text-xl"></i>
+          <span class="ps-4 cursor-pointer text-xl"></span>
         </div>
       </div>
 
+      <div
+        class="aside hidden lg:flex flex-col gap-10 h-full w-full border-r-1 border-gray-300 min-h-dvh text-white ...">
+        <h2 class="text-4xl font-semibold font-logoText2 ps-12 pb-8 mt-10 text-vibPink">Photo</h2>
+        <div class="flex items-center hover:text-darkOrange"><i
+            class="fa-solid fa-house-user cursor-pointer text-xl ps-12"></i> <span
+            class="ps-4 cursor-pointer text-xl">Home</span></div>
+        <div class="flex items-center hover:text-darkOrange"> <i
+            class="fa-solid fa-user cursor-pointer text-xl ps-12"></i>
+          <span class="ps-4 cursor-pointer text-xl">Profile</span>
+        </div>
+
+        <div class="flex items-center hover:text-darkOrange"> <i
+            class="fa-solid fa-people-group cursor-pointer text-xl ps-12"></i> <span
+            class="ps-4 cursor-pointer text-xl">Explore</span> </div>
+        <div class="flex items-center hover:text-darkOrange"> <i
+            class="fa-solid fa-plus cursor-pointer text-xl ps-12"></i>
+          <span class="ps-4 cursor-pointer text-xl">Create</span>
+        </div>
+        <div class="flex items-center hover:text-darkOrange"> <i
+            class="fa-solid fa-bars cursor-pointer text-xl ps-12"></i>
+          <span class="ps-4 cursor-pointer text-xl">More</span>
+        </div>
+      </div>
+
+
       <!-- imp! Right side container -->
 
-      <div class="aside grid grid-rows-4 h-dvh w-full pt-10 place-items-center col-span-4 overflow-y-scroll ...">
+      <div
+        class="aside grid grid-rows-4 col-span-5 h-dvh w-full pt-10 px-5 place-items-center overflow-y-scroll ... bg-black s:px-10 lg:col-span-4 lg:px-0">
 
         <div class="top-container flex flex-col mt-30 items-center">
-          <div class="top flex justify-center gap-10 pb-12 max-w-3xl border-b-1 border-gray-600 ...">
+          <div class="top flex justify-center gap-10 pb-5 max-w-3xl border-b-1 border-gray-600 ... md:pb-12">
 
             <div class="text-center">
-              <div class="w-40 h-40 rounded-full overflow-hidden border-4 border-blue-500 mx-auto">
-                <img id="profile-img" src="https://via.placeholder.com/150" alt="Profile Picture"
+              <div
+                class="w-25 h-25 rounded-full overflow-hidden border-2 border-blue-500 mx-auto s:border-4 s:w-30 s:h-30 md:w-40 md:h-40">
+                <img id="profile-img" src="/images/default-profile-pic.png" alt="Profile Picture"
                   class="w-full h-full object-cover">
               </div>
 
@@ -69,7 +92,6 @@
               </label>
               <input type="file" id="file-input" class="hidden" accept="image/*">
 
-              <p class="mt-2 text-gray-500">Click to upload a new photo.</p>
             </div>
 
             <div class="text-div">
@@ -100,12 +122,12 @@
               </div>
 
 
-              <div class="post-followers-following flex gap-10">
-                <span class="text-white font-medium"><strong>Posts:</strong> 100</span>
-                <span class="text-white font-medium"><strong>Followers:</strong> 100</span>
-                <span class="text-white font-medium"><strong>Following:</strong> 100</span>
+              <div class="post-followers-following flex flex-wrap gap-3 md:gap-10">
+                <span class="text-white font-medium text-s md:text-lg"><strong>Posts:</strong> 100</span>
+                <span class="text-white font-medium text-s md:text-lg"><strong>Followers:</strong> 100</span>
+                <span class="text-white font-medium text-s md:text-lg"><strong>Following:</strong> 100</span>
               </div>
-              <div class="bio text-white font-light pt-12">Actor / Producer / Running in movies since 1981</div>
+              <div class="bio text-white font-light pt-8 md:pt-12">Actor / Producer / Running in movies since 1981</div>
             </div>
           </div>
 
@@ -125,7 +147,7 @@
 
         <!-- imp! posts -->
         <div id="js-grid"
-          class="body-content grid place-items-center grid-cols-3 gap-1 max-w-6xl mt-65 h-full text-9xl text-white w-fullrow-span-4">
+          class="body-content grid place-items-center grid-cols-3 min-w-1 gap-1 max-w-6xl mt-50 h-full text-9xl text-white w-fullrow-span-4 md:mt-65">
         </div>
       </div>
